+++ src/command_modules/azure-cli-iot/azure/cli/command_modules/iot/tests/latest/test_iot_commands.py
@@ -41,6 +41,11 @@ class IoTHubTest(ScenarioTest):
             self.check_pattern('connectionString', conn_str_pattern)
         ])
 
+        self.cmd('iot hub show-connection-string -n {0} -g {1} --all'.format(hub, rg), checks=[
+            self.check('length(connectionString[*])', 5),
+            self.check_pattern('connectionString[0]', conn_str_pattern)
+        ])
+
         # Test 'az iot hub update'
         property_to_update = 'properties.operationsMonitoringProperties.events.DeviceTelemetry'
         updated_value = 'Error, Information'
@@ -145,7 +150,7 @@ class IoTHubTest(ScenarioTest):
             self.check('[0].name', 'TotalMessages'),
             self.check('[0].maxValue', 400000),
             self.check('[1].name', 'TotalDeviceCount'),
-            self.check('[1].maxValue', 1000000)
+            self.check('[1].maxValue', 'Unlimited')
         ])
 
         # Test 'az iot hub show-stats'